+++ public/js/bookings.js
@@ -19,3 +19,20 @@ changeTab(document.getElementsByClassName('tab-title')[0], 'completed')
 function payNow(orderid) {
     window.location.href = `/booking/pay?id=${orderid}`
 }
+
+async function cancelBooking(orderid) {
+    const res = await fetch(`/api/booking/cancel-order?orderid=${orderid}`, {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+    })
+    const data = await res.json();
+
+    if (res.status == 400 || res.status == 401 ) {
+        alert("Error cancelling booking");
+        return;
+    }
+    
+    if (res.status == 200) {
+        window.location.reload();
+    }
+}