+++ routes/api/booking.js
@@ -5,6 +5,7 @@ const { orderCreate } = require('../../models/functions/booking/order-create.js'
 const { transactionComplete } = require('../../models/functions/booking/transaction-complete.js')
 const { orderComplete } = require('../../models/functions/booking/order-complete.js')
 const { carBookDates } = require('../../models/functions/booking/book-car-dates.js')
+const { orderCancel } = require('../../models/functions/booking/cancel-order.js')
 
 // ! Create Order API Endpoint
 router.post('/order-create', authCheck, orderCreate);
@@ -17,4 +18,7 @@ router.post('/transaction-complete', authCheck, transactionComplete)
 
 // ! Book Car Dates after payment completion
 router.post('/book-car-dates', authCheck, carBookDates)
+
+// ! Cancel Order and Cancel Transaction
+router.post('/cancel-order', authCheck, orderCancel);
 module.exports = router;